Fix resourceTypes Methods Incorrectly Merging
With a given `resourceType` the methods where not properly being
merged together.  For instance, a resource declares to be of type
`foo`, and that type defines under it's `get` method two query
parameters of `bar` and `baz`.  If the resource itself also wants
to fold in several more query parameters under the `get` method
they will not be folded in, but instead completely override the
entire `get` method described by the resource.

Looking at the following topic under the raml spec:
https://github.com/raml-org/raml-spec/blob/master/raml-0.8.md#applying-resource-types-and-traits

It _appears_ that the the methods should be merged together and
note simply overriden.